Adjust autocomplete focus boost
This lowers the autocomplete focus boost from 40 to 15. The idea is
that, because the boost for population is 20, the focus can't possibly
override a popular city, but it can come close.

On our default dev build, no acceptance tests fail but the san francisco
autocomplete test now passes!

It's not clear if this fixes WOF venue issues yet, since the latest
build there failed :(